requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-surprise==1.1.4
scipy==1.14.1
six==1.17.0
tzdata==2024.2
Werkzeug==3.1.3
